feat: formulários por serviço em _forms/, checkbox no wizard, campos revisados
